requirements.txt
amqp==5.1.1
asgiref==3.6.0
beautifulsoup4==4.12.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cryptography==40.0.1
defusedxml==0.7.1
Django==4.1.7
django-bootstrap4==22.3
django-cleanup==7.0.0
django-cors-headers==3.14.0
django-environ==0.10.0
django-filter==23.1
django-precise-bbcode==1.2.15
django-ranged-response==0.2.0
django-sass-processor==1.2.2
django-simple-captcha==0.5.17
djangorestframework==3.14.0
easy-thumbnails==2.8.5
gevent==22.10.2
greenlet==2.0.2
idna==3.4
kombu==5.2.4
mysqlclient==2.1.1
oauthlib==3.2.2
Pillow==9.4.0
prompt-toolkit==3.0.38
pycparser==2.21
PyJWT==2.6.0
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.1.0
social-auth-core==4.4.0
soupsieve==2.4
sqlparse==0.4.3
tzdata==2023.3
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
zope.event==4.6
zope.interface==6.0

selenium~=4.8.3
channels~=4.0.0